% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars20.R
\docType{data}
\name{cars_over_20}
\alias{cars_over_20}
\title{Descriptive data title}
\format{a \code{data.frame} containing the following fields:
\describe{
\item{speed}{}
\item{dist}{}
}}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{mtcars20}
}
