% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars20.R
\docType{package}
\name{mtcars20}
\alias{mtcars20}
\alias{mtcars20-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
mtcars20
A data package for mtcars20.

Use \code{data(package='mtcars20')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{cars_over_20}
}
